xml/parsing/manakai/condsects-2.dat

#data
<!DOCTYPE a SYSTEM "http://a/"[
<!ENTITY % a "">
]><a>&x;&y;</a>
#resource http://a/
<![%a;INCLUDE[
  <!ENTITY x "X">
]]>
  <!ENTITY y "Y">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]2;3;xml:dtd:ext decl;;;w
[1]4;3;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   "XY"

#data
<!DOCTYPE a SYSTEM "http://a/"[
<!ENTITY % a "">
]><a>&x;&y;</a>
#resource http://a/
<![ %a;INCLUDE[
  <!ENTITY x "X">
]]>
  <!ENTITY y "Y">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]2;3;xml:dtd:ext decl;;;w
[1]4;3;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   "XY"

#data
<!DOCTYPE a SYSTEM "http://a/"[
<!ENTITY % a "">
]><a>&x;&y;</a>
#resource http://a/
<![INCLUDE%a;[
  <!ENTITY x "X">
]]>
  <!ENTITY y "Y">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]2;3;xml:dtd:ext decl;;;w
[1]4;3;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   "XY"

#data
<!DOCTYPE a SYSTEM "http://a/"[
<!ENTITY % a "">
]><a>&x;&y;</a>
#resource http://a/
<![INCLUDE %a;[
  <!ENTITY x "X">
]]>
  <!ENTITY y "Y">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]2;3;xml:dtd:ext decl;;;w
[1]4;3;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   "XY"

#data
<!DOCTYPE a SYSTEM "http://a/"[
<!ENTITY % a "">
]><a>&x;&y;</a>
#resource http://a/
<![
%a;
INCLUDE
%a;
[
  <!ENTITY x "X">
]]>
  <!ENTITY y "Y">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]6;3;xml:dtd:ext decl;;;w
[1]8;3;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   "XY"

#data
<!DOCTYPE a SYSTEM "http://a/"[
<!ENTITY % a "INCLUDE">
]><a>&x;&y;</a>
#resource http://a/
<![%a;[
  <!ENTITY x "X">
]]>
  <!ENTITY y "Y">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]2;3;xml:dtd:ext decl;;;w
[1]4;3;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   "XY"

#data
<!DOCTYPE a SYSTEM "http://a/"[
<!ENTITY % a "  INCLUDE ">
]><a>&x;&y;</a>
#resource http://a/
<![%a;[
  <!ENTITY x "X">
]]>
  <!ENTITY y "Y">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]2;3;xml:dtd:ext decl;;;w
[1]4;3;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   "XY"

#data
<!DOCTYPE a SYSTEM "http://a/"[
<!ENTITY % a "IGNORE">
]><a>&x;&y;</a>
#resource http://a/
<![%a;[
  <!ENTITY x "X">
]]>
  <!ENTITY y "Y">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]4;3;xml:dtd:ext decl;;;w
3;6;entity not declared;;x;;m
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   "&x;Y"

#data
<!DOCTYPE a SYSTEM "http://a/"[
]><a>&x;&y;</a>
#resource http://a/
<!ENTITY % b "">
<!ENTITY % a "%b;IGNORE%b;">
<![%a;[
  <!ENTITY x "X">
]]>
  <!ENTITY y "Y">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
[1]6;3;xml:dtd:ext decl;;;w
2;6;entity not declared;;x;;m
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   "&x;Y"

#data
<!DOCTYPE a SYSTEM "http://a/"[
]><a>&x;&y;</a>
#resource http://a/
<!ENTITY % a "INC">
<![%a;LUDE[
  <!ENTITY x "X">
]]>
  <!ENTITY y "Y">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;11;ms:no dso;;;m
[1]5;3;xml:dtd:ext decl;;;w
2;6;entity not declared;;x;;m
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   "&x;Y"

#data
<!DOCTYPE a SYSTEM "http://a/"[
]><a>&x;&y;</a>
#resource http://a/
<!ENTITY % a "C">
<![IN%a;LUDE[
  <!ENTITY x "X">
]]>
  <!ENTITY y "Y">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;6;ms:no dso;;;m
[1]5;3;xml:dtd:ext decl;;;w
2;6;entity not declared;;x;;m
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   "&x;Y"

#data
<!DOCTYPE a SYSTEM "http://a/"[
]><a>&x;&y;</a>
#resource http://a/
<!ENTITY % a "INCLUDE[">
<![%a;
  <!ENTITY x "X">
]]>
  <!ENTITY y "Y">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]1;22;mdc in pe in md;;;m
2;6;entity not declared;;x;;m
[1]5;3;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   "&x;Y"

#data
<!DOCTYPE a SYSTEM "http://a/"[
<!ENTITY % a "<![INCLUDE[">
]><a>&x;&y;</a>
#resource http://a/
%a;
  <!ENTITY x "X">
]]>
  <!ENTITY y "Y">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]2;3;xml:dtd:ext decl;;;w
2;26;ms:unclosed;;;m
[1]3;1;string in internal subset;;;m
[1]4;3;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   "XY"

#data
<!DOCTYPE a SYSTEM "http://a/"[
<!ENTITY % a "<![IGNORE [">
]><a>&x;&y;</a>
#resource http://a/
%a;
  <!ENTITY x "X">
]]>
  <!ENTITY y "Y">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]2;3;xml:dtd:ext decl;;;w
2;26;ms:unclosed;;;m
[1]3;1;string in internal subset;;;m
[1]4;3;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   "XY"

#data
<!DOCTYPE a SYSTEM "http://a/"[
<!ENTITY % a SYSTEM "http://b/">
]><a>&x;&y;</a>
#resource http://b/
<![IGNORE[
#resource http://a/
%a;
  <!ENTITY x "X">
]]>
  <!ENTITY y "Y">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[2]1;1;no XML decl;;;s
[1]1;11;ms:unclosed;;;m
[2]2;3;xml:dtd:ext decl;;;w
[2]3;1;string in internal subset;;;m
[2]4;3;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   "XY"
